model, cli: ModelOptions alone sets the f16 cache default, and the split check takes a cache type
The CLI's parameters said f16 while ModelOptions said f32, so a model built without the CLI's flags stored f32 caches.
The synthetic bench, which the perf gate runs, and llmx-split-check, the layer split's bit-identity gate, therefore never ran the caches that users and the server run.
ModelOptions now holds the one default, and the CLI starts from it, changes a side only when --cache-type-k or --cache-type-v is given, and prints the default from it in the help.
The two cache type names move beside KVType, so the CLI and the split check read them the same way.
The split check takes a cache type after the ubatch, f16 unless f32 is given, so its gate covers both types.
The kv-cache transaction check, which counts cache bytes as floats, and the matched CPU comparison, whose other arm stores f32, ask for f32 sides.
The tests that compare two models built with the same options keep the default and now run f16.

+++ src/model/arch_qwen.hpp
@@ -197,9 +197,10 @@ struct Placement {
 };
 
 // Choices made once at construction, before the caches are allocated: how each cache side is stored (backend.hpp KVType, the CLI's --cache-type-k and --cache-type-v), the same on every backend or refused.
+// The runtime's default cache type is set here and nowhere else: the CLI changes a side only when its flag is given.
 struct ModelOptions {
-    backend::KVType kv_k = backend::KVType::f32;
-    backend::KVType kv_v = backend::KVType::f32;
+    backend::KVType kv_k = backend::KVType::f16;
+    backend::KVType kv_v = backend::KVType::f16;
     // Tokens the KV pool holds in total, shared by every sequence; zero means one model context, which is what one conversation needs and what a server divides among its requests unless told otherwise.
     size_t kv_tokens = 0;
 };
